+++ packages/connect/src/components/SequenceConnectInlineProvider/SequenceConnectInlineProvider.tsx
@@ -38,7 +38,7 @@ import { EpicAuthProvider } from '../EpicAuthProvider/index.js'
 import { JsonTreeViewer } from '../JsonTreeViewer.js'
 import { NetworkBadge } from '../NetworkBadge/index.js'
 import { PageHeading } from '../PageHeading/index.js'
-import { PoweredBySequence } from '../SequenceLogo/index.js'
+import { PoweredByPolygon } from '../PolygonBrand/index.js'
 import { ShadowRoot } from '../ShadowRoot/index.js'
 import { SocialLink } from '../SocialLink/SocialLink.js'
 import { TxnDetails } from '../TxnDetails/index.js'
@@ -379,7 +379,7 @@ export const SequenceConnectInlineProvider = (props: SequenceConnectInlineProvid
                                 </div>
 
                                 <div className="mt-4">
-                                  <PoweredBySequence />
+                                  <PoweredByPolygon />
                                 </div>
                               </div>
                             </Modal>