+++ packages/hooks/src/hooks/IndexerGateway/useGetNativeTokenBalance.ts
@@ -2,7 +2,7 @@ import { SequenceIndexerGateway, type IndexerGateway, type TokenBalance } from '
 import { useQuery } from '@tanstack/react-query'
 
 import { QUERY_KEYS, time } from '../../constants.js'
-import type { HooksOptions } from '../../types/hooks.js'
+import type { QueryHookOptions } from '../../types/hooks.js'
 import { createNativeTokenBalance } from '../../utils/helpers.js'
 
 import { useIndexerGatewayClient } from './useIndexerGatewayClient.js'
@@ -60,14 +60,15 @@ const getNativeTokenBalance = async (
  * }
  * ```
  */
-export const useGetNativeTokenBalance = (args: GetNativeTokenBalanceArgs, options?: HooksOptions) => {
+export const useGetNativeTokenBalance = (args: GetNativeTokenBalanceArgs, options?: QueryHookOptions<TokenBalance[]>) => {
   const indexerGatewayClient = useIndexerGatewayClient()
 
   return useQuery({
-    queryKey: [QUERY_KEYS.useGetNativeTokenBalance, args, options],
+    queryKey: [QUERY_KEYS.useGetNativeTokenBalance, args],
     queryFn: async () => await getNativeTokenBalance(indexerGatewayClient, args),
-    retry: options?.retry ?? false,
+    retry: false,
     staleTime: time.oneSecond * 30,
-    enabled: !!args.accountAddress && !options?.disabled
+    enabled: !!args.accountAddress,
+    ...options
   })
 }