+++ packages/hooks/src/types/hooks.ts
@@ -1,4 +1,9 @@
-export interface HooksOptions {
-  disabled?: boolean
-  retry?: boolean | number
-}
+import type { InfiniteData, QueryKey, UseInfiniteQueryOptions, UseQueryOptions } from '@tanstack/react-query'
+
+export type QueryHookOptions<TQueryFnData = unknown, TError = Error, TData = TQueryFnData> = Partial<
+  UseQueryOptions<TQueryFnData, TError, TData, QueryKey>
+>
+
+export type InfiniteQueryHookOptions<TQueryFnData = unknown, TError = Error, TPageParam = unknown> = Partial<
+  UseInfiniteQueryOptions<TQueryFnData, TError, InfiniteData<TQueryFnData>, QueryKey, TPageParam>
+>