+++ packages/wallet-widget/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsequence/kit-wallet
 
+## 6.0.6
+
+### Patch Changes
+
+- Minor improvements
+- Updated dependencies
+  - @0xsequence/connect@6.0.6
+  - @0xsequence/hooks@6.0.6
+
+## 6.0.5
+
+### Patch Changes
+
+- Powered by logo update
+- Updated dependencies
+  - @0xsequence/connect@6.0.5
+  - @0xsequence/hooks@6.0.5
+
 ## 6.0.4
 
 ### Patch Changes